fixed some compilation bugs
git-svn-id: svn://svn.gna.org/svn/freedoko/trunk@1308 d9929be3-8b05-0410-8f9e-ffa92982859d

+++ src/player/ai/cards_information.heuristics.cpp
@@ -599,7 +599,8 @@ namespace CardsInformationHeuristics {
         if (played_card == card->first)
           break;
         else if (former_winnercard.less(card->first))
-          CHANGE_WEIGHTING(card->first, weighting * card->second);
+          CHANGE_WEIGHTING(card->first,
+                           static_cast<int>(weighting * card->second));
       }
 
       return ;
